+++ src/Collections/Collection.php
@@ -36,7 +36,7 @@ class Collection implements CollectionInterface
      * @param string|null $type  What type of items the collection will hold.
      *     Either a value returned by gettype() or a fully qualified class name.
      */
-    public function __construct(Wye $wye, $items = [], $type = null)
+    public function __construct(Wye $wye, array $items = [], $type = null)
     {
         $this->setWye($wye);
 
@@ -228,6 +228,21 @@ class Collection implements CollectionInterface
         }
 
         return $default;
+    }
+
+    /**
+     * Run a map over each of the items.
+     *
+     * @param  callable  $callback
+     * @return static
+     */
+    public function map(callable $callback)
+    {
+        $keys = array_keys($this->items);
+
+        $items = array_map($callback, $this->items, $keys);
+
+        return new static($this->getWye(), array_combine($keys, $items));
     }
 
     /**