+++ tests/Unit/Wye/ResetTest.php
@@ -71,4 +71,48 @@ class ResetTest extends TestCase
 
         $this->assertAttributeSame([], "transactions", Wye::class);
     }
+
+    public function testBacktraceSingleResetToFalse()
+    {
+        Wye::logBacktraceForTest();
+
+        $this->assertTrue(Wye::getBacktraceSingle());
+
+        Wye::reset();
+
+        $this->assertFalse(Wye::getBacktraceSingle());
+    }
+
+    public function testBacktraceAllNotResetToFalse()
+    {
+        Wye::logBacktraceForAllTests();
+
+        $this->assertTrue(Wye::getBacktraceAll());
+
+        Wye::reset();
+
+        $this->assertTrue(Wye::getBacktraceAll());
+    }
+
+    public function testBacktraceLimitResetToNull()
+    {
+        Wye::setBacktraceLimit(12);
+
+        $this->assertSame(12, Wye::getBacktraceLimit());
+
+        Wye::reset();
+
+        $this->assertNull(Wye::getBacktraceLimit());
+    }
+
+    public function testBacktraceDefaultLimitNotReset()
+    {
+        Wye::setBacktraceDefaultLimit(12);
+
+        $this->assertSame(12, Wye::getBacktraceDefaultLimit());
+
+        Wye::reset();
+
+        $this->assertSame(12, Wye::getBacktraceDefaultLimit());
+    }
 }